// sourcecode/BTrak.Packages/snova-billing-widgets/billing-module-components/billing-components/src/lib/billing-widgets/models/groupE.model.ts
export class GroupE {
    id: string;
    siteId: string;
    grdId: string;
    bankId: string;
    grdName: string;
    bankAccountName: string;
    email: string;
    siteName: string;
    month: Date;
    startDate: Date;
    endDate: Date;
    term: string;
    year: Date;
    production: number;
    reprise: number;
    facturation: number;
    autocTariff: number;
    gridInvoice: string;
    gridInvoiceName: string;
    solarLogValueKwh: number;
    productionPercent: number;
    plannedAutoC: number;
    plannedSystem: number;
    invoiceUrl: string;
    gridInvoiceDate: Date;
    isGre: boolean;
    hauteTariff: number;
    basTariff: number;
    distribution: number;
    greFacturation: number;
    administrationRomandeE: number;
    autoConsumptionSum : number;
    facturationSum : number;
    autoConsumption: number;
    tariffTotal: number;
    greTotal: number;
    subTotal: number;
    tva: number;
    tvaForSubTotal: number;
    total: number;
    confirmDetailsfromGrid: boolean;
    autoCTariff: number;
    generateInvoice: boolean;
    isArchived: boolean;
    timeStamp: string;
    pRATotal: any;
    pRAFields: string;
    dFFields: string;
    totalValue: any;
    outStandingAmount: number;
    messageType: string;
}